requirements.txt
alpaca-py==0.15.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
billiard==4.2.2
cachetools==6.2.1
celery==5.3.4
certifi==2025.10.5
channels==4.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
google-auth==2.42.1
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
kombu==5.5.4
msgpack==1.1.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==10.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg[binary]
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.1
requests==2.31.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sseclient-py==1.8.0
stripe==7.8.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.7.0
websockets==11.0.3
wheel==0.45.1
whitenoise==6.11.0
stripe==5.4.0
python-dotenv==1.0.0